[WIP] Add `ParallelIterator::fold_unordered`
On the surface, this looks like `fold`, but it does **not** preserve the
relative order among folded items, so it's only suitable for commutative
operations. The advantage is that it can better reuse the accumulators
even across work-stealing. The implementation is fairly simplistic with
`Mutex` locks, but hopefully those will be mostly uncontended.